Fixed bug creating datachannel (wrong type of period)

+++ datachannel.h
@@ -93,7 +93,7 @@ public:
     int dcmode;
     int ready;
     int state;
-    int period;
+    unsigned long period;
     int edge;
 
     //trigger controlling variables
